Drop support for PyCrypto - M2Crypto is now a hard requirement
There were some security flaws in the way we used PyCrypto. We could have changed our use of PyCrypto to be more secure, but it seems more efficient to drop support for PyCrypto rather than maintain two different crypto backends. M2Crypto has more batteries included, so using it instead of PyCrypto is more in keeping with the "leave security to the experts" approach.

Fixes #11 and #12.